Assign Permissions to Role Functionality Added

+++ app/Http/Controllers/RoleController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Permission;
 use App\Models\Role;
 use Illuminate\Http\Request;
 
@@ -51,14 +52,46 @@ class RoleController extends Controller
             'name' => ['required', 'min:4', 'unique:roles,name,' . $role->id]
         ]);
 
-        $role->update($valid);
+        if (!auth()->user()->hasRole($role)) {
+            $role->update($valid);
 
-        return redirect()->route('roles.index');
+            return redirect()->route('roles.index');
+        }
+        else
+            return back();
     }
 
     public function destroy(Role $role)
     {
-        $role->delete();
+        if (!auth()->user()->hasRole($role)) {
+            $role->delete();
+
+            return redirect()->route('roles.index');
+        }
+        else
+            return back()->with('error', "Can't delete your Role");
+    }
+
+    public function assignPermissionsForm(Role $role)
+    {
+        $permissions = Permission::all();
+
+        return inertia('Roles/AssignPermission', [
+            'role' => $role,
+            'permissions' => $permissions,
+            'assigned' => $role->permissions
+        ]);
+    }
+
+    public function assignPermissions(Request $request, Role $role)
+    {
+        $permissions = $request->input('permissions');
+
+        $role->permissions()->detach();
+
+        foreach ($permissions as $key => $permission) {
+            $role->permissions()->sync($permissions[$key], false);
+        }
 
         return redirect()->route('roles.index');
     }
